+++ src/cray/crinput.h
