+++ src/cray/crprofile.c
@@ -1,0 +1,1 @@
+#include "crprofile.h"
