+++ src/cray/crprofile.h
@@ -1,0 +1,10 @@
+#ifndef _CRAY_PROFILE_H_
+#define _CRAY_PROFILE_H_
+
+#include <stdint.h>
+
+typedef struct CycleProfile {
+	uint64_t totalRenderTime;
+} CycleProfile;
+
+#endif // !_CRAY_PROFILE_H_
