+++ src/cray/crtime.h
@@ -1,4 +1,5 @@
-#ifndef __CRAY_TIME_H_
+#ifndef _CRAY_TIME_H_
+#define _CRAY_TIME_H_
 
 #include <stdint.h>
 #include <float.h>
@@ -8,4 +9,4 @@ extern double GetTimeInS(uint64_t elapsedTicks);
 extern double GetTimeInMS(uint64_t elapsedTicks);
 extern double GetTimeInUS(uint64_t elapsedTicks);
 
-#endif // !__CRAY_TIME_H_
+#endif // !_CRAY_TIME_H_
